# src/server/api-doc.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "JACC APIs"
host: "jacc-api.herokuapp.com"
basePath: "/api"
schemes:
- "https"
- "http"
paths:
...
  /admin/medicine/insert:
    post:
      tags:
      - "Medicine"
      summary: "Add medicine"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          properties:
            name:
              type: string
            mg:
              type: integer
            qty:
              type: integer
            unit_price:
              type: integer
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
  /admin/medicine/update:
    post:
      tags:
      - "Medicine"
      summary: "Update medicine"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          properties:
            name:
              type: string
            mg:
              type: integer
            qty:
              type: integer
            unit_price:
              type: integer
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
  /admin/medicine/get-total-sale/today:
    get:
      tags:
      - "Medicine"
      summary: "Get total sales for this day"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
...
  /admin/patient/insert:
    post:
      tags:
      - "Patient"
      summary: "Insert Patient"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
  /admin/patient/get-list?name=test&page=2&order=asc:
    get:
      tags:
      - "Patient"
      summary: "Get Patient List"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
  /admin/patient/get-info/:patient_id:
    get:
      tags:
      - "Patient"
      summary: "Get Patient Info"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
...
  /admin/medicine/get-list?name=test&page=2&order=asc:
    get:
      tags:
      - "Medicine"
      summary: "Get Medicine List"
      description: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
...